Load Transaction data automatically based on the current selected month
Using MomentJs

Colin's comments about naming.

Remove unnecessary braces.

Load transactions as a side effect of changing selected month. Always make
sure that month is fully loaded when the tab is selected. Currently the
list of transactions is not filtered by month. Transactions still loaded in
pages so as to not waste http requests.