Refactor of businessDetails (#96)
* Basic trader screen with no business details

Send payment button is fixed to the bottom of the screen
All transactions (not filtered to specific trader yet) show when user is logged in

* Start to populate the business details

BusinessDetails.js has been renamed to TraderScreen.js because it is the skeleton for the Trader screen which contains much more than just business details

BusinessDetails.js now contains the code for page of theheader of Trader screen, which contains the entries of attributes of the business eg. address, opening times etc.

* WIP

* Add images and fields to the trader screen

* Refactor trader modal

trader modal has now been reduxified
Additional api calls to get extra information - call this the "business profile"and merge it in

Please enter the commit message for your changes. Lines starting

* Rename businessDetails screen to traderScreen

* Fix conflicts

* Address code review comments

* extend logic

* correct releative paths

* minor correction