Refactor TraderScreen (#171)
* Add general ProfileScreen component

To handle both TraderScreen and PersonScreen (when it is implemented). This is a refactor of TraderScreen. This commit contains just moving files around and initial planned structure for TraderScreen

* Refactor CategoryImage

Before: CategoryImage was used for placeholder images
Now: ProfileImage used for both placeholder images and actual images so that render code for the TraderScreen (ProfileScreen) will just be one line, with no if statement. Logic has been moved inside. ProfileImage is hence more general.

* ProfileScreen is now a stateless component

* Move close and expand icons to profileScreen

Now there is only one file in traderScreen so move file down a level and delete folder

* Use ProfileScreen and ProfileImage everywhere, and fix imports etc.

* Further tidy up ProfileImage

* Fix imports and exports